ADH-7557 Pull changes from PR
